[PSL-844] Change calculation of registration/pre-burn fees for Sense and NFT (#182)
* [PSL-844] Change calculation of registration/pre-burn fees for Sense and NFT

- introduced "global fee adjustment multiplier" which is used in calculation of reg fees
- changed default masternode fees (without adjustments):
    - default storage fee per MB: 5000 PSL
	- default ticket blockchain storage fee per KB: 200 PSL
	- sense processing fee per MB: 50 PSL
	- sense compute fee: 5000 PSL
- changed AVERAGE_SENSE_DUPE_DATA_SIZE_MB: 10 MB
- introduced NFT_DISCOUNT_MULTIPLIER=0.45
- added version support for MasterNode messages (v1):
  - added new masternode fees to announce for masternodes:
    - Sense Compute Fee
	- Sense Processing Fee Per MB
- calculate new fees (Sense Compute Fee, Sense Processing Fee Per MB) as masternode's network median (trimmean with fixed 25%)
- implemented cache for chain deflator factor:
   - deflator factor is changing only every 10000 blocks
   - introducted a map to cache deflator factor for every 10000 blocks (cache key)
   - use RW shared_lock/unique_lock to access map to retrieve deflator factor fast
- changed sense fee calculation:
  [ nStorageFeePerMB * AVERAGE_SENSE_DUPE_DATA_SIZE_MB  +
    nDataSizeInMB * nSenseFeePerMB +
	SENSE_COMPUTE_FEE +
	nTicketFeePerKB * ACTION_SENSE_TICKET_SIZE_KB ] * GLOBAL_FEE_ADJUSTMENT_MULTIPLIER * DEFLATOR_FACTOR
- changed cascade fee calculation:
  [ nStorageFeePerMB * nDataSizeInMB +
    nTicketFeePerKB * ACTION_CASCADE_TICKET_SIZE_KB ] * GLOBAL_FEE_ADJUSTMENT_MULTIPLIER * DEFLATOR_FACTOR
- changed NFT fee calculation:
  [ SenseFee_WithoutTicketFee + Cascade_WithoutTicketFee ] * NFT_DISCOUNT_MULTIPLIER +
    nNftTicketDataSizeInKB * nTicketFeePerKB * GLOBAL_FEE_ADJUSTMENT_MULTIPLIER * DEFLATOR_FACTOR
- enhanced storagefee RPC API:
    - added optional height parameter to get historical fee for the specific height, default height is current chain height
	- storagefee API uses global_fee_adjustment_multiplier and deflator factor
	- added new APIs to retrieve new fees: getsensecomputefee, getsenseprocessingfee
	- refactored storagefee
- storagefee API:
  - setfee <n>                           - Set storage fee for MN.
  - getnetworkfee	(<height>)           - Get Network median storage fee (per MB).
  - getnftticketfee (<height>)           - Get Network median NFT ticket fee (per KB).
  - getsensecomputefee (<height>)        - Get Network median Sense Compute fee.
  - getsenseprocessingfee (<height>)     - Get Network median Sense Processing fee (per MB).
  - getlocalfee (<height>)               - Get local masternode storage fee (per MB).
  - getactionfees <data_size> (<height>) - Get action fee by data size (in MB)

* changed version to 1.2.0.14

* fee API changes

* python tests

* mingw debug build fix

* debug build

* fix

* python tests fixes

* storage-fee api - 0-height

* some refactoring

* changes

* disabled CodeQL workflow

* CFlatDB changes

+++ src/gtest/test_getarg.cpp
@@ -32,7 +32,7 @@ static void ResetArgs(const string& strArg)
     for (const auto& s : vecArg)
         vecChar.push_back(s.c_str());
 
-    ParseParameters(vecChar.size(), &vecChar[0]);
+    ParseParameters(static_cast<int>(vecChar.size()), &vecChar[0]);
 }
 
 TEST(test_getarg, boolarg)
